backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
websockets==10.0
python-multipart==0.0.5
pydantic==1.8.2
sqlalchemy==1.4.31
psycopg2-binary==2.9.3
alembic==1.7.5
python-jose==3.3.0
passlib==1.7.4
